First test added and passing.
The first user story is completed with a simple each block in the view and quality method on the idea. That will eventually be broken out into the idea decorator, but for now it's fine where it is. The next step is to make some models tests to assue the idea behaves correctly.